Refactor out ServiceControlAuditsByBinarySearch
